[Coroutines] Implement elision for noop coroutines
