+++ llvm/test/Transforms/Coroutines/coro-elide-noop.ll
@@ -1,0 +1,13 @@
+; NOTE: Assertions have been autogenerated by utils/update_test_checks.py UTC_ARGS: --version 6
+; Tests that noop coroutine can be elided
+; RUN: opt < %s -S -passes='coro-elide' | FileCheck %s
+
+define void @fn() {
+; CHECK-LABEL: define void @fn() {
+; CHECK-NEXT:    ret void
+;
+  %noop.frame = tail call noundef ptr @llvm.coro.noop()
+  %resume = tail call ptr @llvm.coro.subfn.addr(ptr %noop.frame, i8 0)
+  tail call fastcc void %resume(ptr %noop.frame)
+  ret void
+}
